split the questionnaire modal in sub-components
