+++ package.json
@@ -31,6 +31,7 @@
     "buffer": "^6.0.3",
     "i18n-js": "^3.8.0",
     "lodash.clonedeep": "^4.5.0",
+    "lodash.debounce": "^4.0.8",
     "lodash.memoize": "^4.1.2",
     "node-forge": "^1.2.1",
     "react": "^17.0.2",
